requirements.txt
langchain
langchain_community
langchain_openai
langchain_groq
langchain-ollama
langchain_huggingface
sentence-transformers
langchain_core
streamlit
ipykernel
python-dotenv
arxiv
wikipedia
duckduckgo-search
ddgs
faiss-cpu
faiss-gpu
pypdf
langchain_chroma
mysql
validators
youtube_transcript_api
pytube
numexpr
langchain_experimental
unstructured
boto3
awscli
openai
langchain_nvidia_ai_endpoints